add extra unittest for get_best_available_device()
